DataGot関数を移動
